# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API documentation for authentication, translations, locales, approval, bulk operations, and export.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get paginated translations",
                "operationId": "listTranslations",
                "responses": {
                    "200": {
                        "description": "Paginated translation list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create translation",
                "operationId": "storeTranslation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTranslationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/{translation}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get one translation",
                "operationId": "showTranslation",
                "parameters": [
                    {
                        "name": "translation",
                        "in": "path",
                        "description": "Translation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete translation",
                "operationId": "deleteTranslation",
                "parameters": [
                    {
                        "name": "translation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update translation",
                "operationId": "updateTranslation",
                "parameters": [
                    {
                        "name": "translation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTranslationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/search": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search translations",
                "operationId": "searchTranslations",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchTranslationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranslationListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/bulk": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Bulk create translations",
                "operationId": "bulkCreateTranslations",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkTranslationsRequest"
                            },
                            "examples": {
                                "multiple_translations": {
                                    "summary": "Bulk create multiple translations",
                                    "value": {
                                        "translations": [
                                            {
                                                "key": "home.title",
                                                "locale": "en",
                                                "value": "Home",
                                                "tags": [
                                                    "web"
                                                ]
                                            },
                                            {
                                                "key": "home.title",
                                                "locale": "fr",
                                                "value": "Accueil",
                                                "tags": [
                                                    "web"
                                                ]
                                            },
                                            {
                                                "key": "auth.login.button",
                                                "locale": "en",
                                                "value": "Login",
                                                "tags": [
                                                    "web",
                                                    "mobile"
                                                ]
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bulk created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkTranslationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/{translation}/approve": {
            "patch": {
                "tags": [
                    "Translations"
                ],
                "summary": "Approve translation",
                "operationId": "approveTranslation",
                "parameters": [
                    {
                        "name": "translation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TagResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Web"
                    },
                    "slug": {
                        "type": "string",
                        "example": "web"
                    },
                    "description": {
                        "type": "string",
                        "example": "Web context",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Translation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "key": {
                        "type": "string",
                        "example": "auth.login.title"
                    },
                    "locale": {
                        "type": "string",
                        "example": "en"
                    },
                    "value": {
                        "type": "string",
                        "example": "Login"
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TranslationListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Translation"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 10
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 50
                            },
                            "total": {
                                "type": "integer",
                                "example": 500
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "CreateTranslationRequest": {
                "required": [
                    "key",
                    "locale",
                    "value"
                ],
                "properties": {
                    "key": {
                        "type": "string",
                        "example": "auth.login.title"
                    },
                    "locale": {
                        "type": "string",
                        "example": "en"
                    },
                    "value": {
                        "type": "string",
                        "example": "Login"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "web",
                            "mobile"
                        ]
                    },
                    "key_description": {
                        "type": "string",
                        "example": "Login page title"
                    }
                },
                "type": "object"
            },
            "UpdateTranslationRequest": {
                "required": [
                    "value"
                ],
                "properties": {
                    "value": {
                        "type": "string",
                        "example": "Sign In"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "web"
                        ]
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "SearchTranslationRequest": {
                "required": [
                    "search_type",
                    "query"
                ],
                "properties": {
                    "search_type": {
                        "type": "string",
                        "example": "key",
                        "enum": [
                            "key",
                            "content",
                            "tag"
                        ]
                    },
                    "query": {
                        "type": "string",
                        "example": "auth.login"
                    },
                    "locale": {
                        "type": "string",
                        "example": "en"
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "BulkTranslationsRequest": {
                "required": [
                    "translations"
                ],
                "properties": {
                    "translations": {
                        "type": "array",
                        "items": {
                            "required": [
                                "key",
                                "locale",
                                "value"
                            ],
                            "properties": {
                                "key": {
                                    "type": "string",
                                    "example": "home.title"
                                },
                                "locale": {
                                    "type": "string",
                                    "example": "en"
                                },
                                "value": {
                                    "type": "string",
                                    "example": "Home"
                                },
                                "tags": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": [
                                        "web"
                                    ]
                                },
                                "key_description": {
                                    "type": "string",
                                    "example": "Homepage title"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "BulkTranslationResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Translation"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the Sanctum API token returned by /login. Example: Bearer 1|xxxx",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}